refactor: Moved types staves.core.Libc to the Gentoo builder module.
The goal of this refactoring is that the builders.gentoo module is self-contained. If it is only dependant on the Python standard library, it is possible to copy it into running containers at runtime. This is much faster than compiling a custom builder image and will speed up development significantly.